# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agent Orchestrator API
  description: >
    API for submitting autonomous agent goals, managing runs, scheduling
    recurring tasks, coordinating multi-agent workflows, and exploring websites.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/plugins:
    get:
      tags: [System]
      summary: List registered plugins
      operationId: listPlugins
      responses:
        "200":
          description: Plugin list
          content:
            application/json:
              schema:
                type: object
                properties:
                  plugins:
                    type: array
                    items:
                      type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/knowledge/stats:
    get:
      tags: [System]
      summary: Knowledge base statistics
      operationId: getKnowledgeStats
      responses:
        "200":
          description: Knowledge stats
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── API Keys ─────────────────────────────────────────────────────
  /api/v1/keys:
    post:
      tags: [Keys]
      summary: Create a new API key
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                    description: The generated API key (shown only once)
                  name:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Runs ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        API key for authentication. Disable auth in dev mode by setting
        AGENT_API_AUTH=false.
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...